% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_diffs.R
\name{compare_diffs}
\alias{compare_diffs}
\title{Compare differences between local lesson files and version in lessons package}
\usage{
compare_diffs(
  pkg = "lhLessons",
  org = "nceas-learning-hub",
  pkg_version = NULL,
  branch = "main",
  prefix = "s",
  file_out = "git_diff.txt"
)
}
\arguments{
\item{pkg}{A character name of the lessons package, default \code{lhLessons}.}

\item{org}{Character Github org (or user) where the package lives, default \code{nceas-learning-hub}.}

\item{pkg_version}{Character Version tag for package, default NULL indicates latest.}

\item{branch}{A character name of the branch to be checked.  Currently only checks against \code{main}.}

\item{prefix}{A character indicator of the expected prefix letter for lesson files.}

\item{file_out}{The location in the current repository where the results will
be saved as a text file.}
}
\value{
(invisibly) a data frame containing \code{file_remote}, \code{file_local},
\code{result} (of \verb{git diff}) and \code{status}.  Writes results out to a text file
(default \code{git_diff.txt})
}
\description{
This function uses \verb{git diff} to compare the local versions of lesson to their
counterpart within a given lessons package (default \code{lhLessons}).  Some notes:
}
\details{
\itemize{
\item This compares using the (bare) lesson name (e.g., \code{s02_git_setup.qmd} is compared
to the package lesson named \code{git_setup.qmd}), so if the course designer has changed
a lesson filename manually, that file will NOT be compared, and assumed to be
a new lesson entirely.
\item Changes in auxiliary files, e.g., images, slides, or data, will NOT be compared.
However, if the path to these files change (e.g., new or renamed image), those
changes in the lesson file will be noted.
\item As of now, checks against the \code{main} branch of the currently installed lessons
package version.  User should confirm that the lessons package version is up to date!
}
}
\examples{
\dontrun{
  compare_diffs()
}
}
