% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_metadata.R
\name{get_course_metadata}
\alias{get_course_metadata}
\title{Get course metadata}
\usage{
get_course_metadata()
}
\value{
A named vector of metadata fields and values
}
\description{
Course metadata is set up in init_course, to contain info on the course name,
repo name, organization, description, and course dates.
}
