% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qmd.R
\name{get_qmd}
\alias{get_qmd}
\title{Copy a lesson file from the package to the location}
\usage{
get_qmd(lesson_id, loc = ".")
}
\arguments{
\item{lesson_id}{The bare filename of the lesson (without file extension)}

\item{loc}{The location on the local computer where the file should be copied}
}
\value{
TRUE if the file copied successfully
}
\description{
Copy a lesson file from the package to the location
}
\examples{
# get_qmd('git-github-intro')
}
