% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_course.R
\name{init_course}
\alias{init_course}
\title{Initialize a coreR course}
\usage{
init_course(course_name, loc = ".")
}
\arguments{
\item{course_name}{The name of the course to be created.  This will set up a
new R project with the given name, populate it with some basic structure, and
optionally commit the project as a Github repository.  This will not
create the lesson structure.}

\item{loc}{The file location where the course repository will be created.
Defaults to the current working directory.}
}
\value{
stuff
}
\description{
Initialize a coreR course
}
\examples{
# init_course(course_name = "2024-10-coreR")
}
