% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_theme.R
\name{install_theme}
\alias{install_theme}
\title{Install a Quarto theme}
\usage{
install_theme(
  org = "nceas-learning-hub",
  theme = c("lh", "adc", "delta", "corer")[1]
)
}
\arguments{
\item{org}{Organization repository where theme extension is located.
Default: \code{nceas-learning-hub}.  Extension repo is called \code{theme_X} where \code{X} is
one of the available themes.}

\item{theme}{The name of the theme within the \verb{_extensions} folder in the repository.
Currently, one of \code{lh}, \code{adc}, \code{delta}, \code{corer}.}
}
\value{
If successful, returns the extension location (org/theme repo)
}
\description{
Install or update a desired Quarto theme to be applied to the course site.
Defaults to \code{nceas-learning-hub/lh_theme}.  This is called in \code{setup_lessons()}
but if the theme changes down the line, this can be used to update it by pulling
from an updated repo or a different repo.
}
\details{
Note this can also be done manually in the Terminal, using \verb{quarto add <org name>/<theme extension repo>}.
}
