% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_course.R
\name{publish_course}
\alias{publish_course}
\title{Publish course}
\usage{
publish_course()
}
\description{
After all the various files are set up, and the course repo is set up with .git
and connected to GitHub, it is ready to publish.  The GitHub Action for publishing
is already installed during \code{setup_course_structure()}, but this final command
checks to make sure .git and GitHub are enabled, then calls \verb{quarto publish gh-pages --no-prompt}.
Note, this could also be done manually in the Terminal.
}
\examples{
\dontrun{publish_course()}
}
