% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_lessons_tags.R
\name{query_lessons_tags}
\alias{query_lessons_tags}
\title{Query the tags of the lessons package remote}
\usage{
query_lessons_tags(
  pkg,
  org = "nceas-learning-hub",
  latest = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{pkg}{The name of the lessons package repository (e.g., "lhLessons")}

\item{org}{The GitHub organization where the lessons package repo lives.
Default "nceas-learning-hub"}

\item{latest}{Return only the latest tag?  Default \code{FALSE} returns all tags.}

\item{quiet}{Suppress display of the version as a message?}
}
\value{
Vector of tags for the package, as characters.
}
\description{
Query the tags of the lessons package remote
}
\examples{
\dontrun{query_lessons_tags(pkg = 'lhLessons', latest = TRUE)}

}
