% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_course.R
\name{setup_course}
\alias{setup_course}
\title{Set up a coreR course}
\usage{
setup_course(course_name)
}
\arguments{
\item{course_name}{The name of the course to be created.  This will set up a
new R project with the given name, populate it with the content, and
commit the project as a Github repository.}
}
\value{
stuff
}
\description{
Set up a coreR course
}
\examples{
setup_course(course_name = '2024-10-coreR')
}
