% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lessons.R
\name{setup_lessons}
\alias{setup_lessons}
\title{Set up a schedule of course lessons}
\usage{
setup_lessons(lessons, overwrite = FALSE)
}
\arguments{
\item{lessons}{A character vector containing the names (bare or .qmd extension) of the lessons, in order.}

\item{overwrite}{If there is already an existing lessons structure, and overwrite == FALSE, will return an error.  If the user wishes to create a new lesson structure, then overwrite == FALSE will remove the existing structure and create a new structure based on the lesson names.}
}
\description{
Create the structure necessary to implement an ordered set of lessons in the Quarto book format.
The user provides a character vector of lesson names (filenames, bare or with \code{.qmd} extension);
these and associated materials (images, data) are copied from \code{coreRlessons} package into the
course project (created with \code{init_course()}).  Sets up the index.qmd, _quarto.yml, and some
additional files.
}
\examples{
\dontrun{setup_lessons(lessons = c('github_introduction', 'r_functions.qmd')}
}
