% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_course_repo.R
\name{verify_course_repo}
\alias{verify_course_repo}
\title{Verify that the current working directory is the correct course!}
\usage{
verify_course_repo(query = "Is this the correct course location?")
}
\arguments{
\item{query}{A string to prompt the user with, generally related to the action
being verified, e.g., 'Do you want to continue with this action in this location?'}
}
\description{
Verify that the current working directory is the correct course!
}
