% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkin_lessons.R
\name{checkin_lessons}
\alias{checkin_lessons}
\title{Check in modified lessons back into lessons package repository}
\usage{
checkin_lessons(
  lessons = NULL,
  pkg = "lhLessons",
  org = "nceas-learning-hub",
  branch = NULL
)
}
\arguments{
\item{lessons}{A character vector or dataframe of lessons to be checked in}

\item{pkg}{The package that contains the lessons; here this represents the
repository name for the package.  Default "lhLessons"}

\item{org}{The GitHub organization where the lessons package repo lives.  Default "nceas-learning-hub"}

\item{branch}{Optional, a branch name to push lesson changes to.  Default \code{NULL}
will result in the course project name (contained in the \code{metadata_course.csv} file) being
used as the branch name}
}
\description{
This function creates a new branch of the given lessons repository and copies
new or updated lessons over to the repository.  Branch name defaults to the
course name but can be overwritten if desired.
}
\details{
Lessons can be provided as a character vector of the lesson file names (with or without
the prefix/extension ornamentation), or as a dataframe resulting from a call to
\code{compare_diffs()}.  Files copied include the lesson .qmd files, plus any associated
image/data/slides folders.

Commits then pushes changes back to Github
}
