% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lessons_version.R
\name{get_lessons_version}
\alias{get_lessons_version}
\title{Query the version of the lessons package}
\usage{
get_lessons_version(pkg = "coreRlessons", quiet = FALSE)
}
\arguments{
\item{pkg}{The name of the lessons package (e.g., "coreRlessons")}
}
\value{
The version of the package, as a character.
}
\description{
Query the version of the lessons package
}
\examples{
\dontrun{get_lessons_version(pkg = "coreRlessons")}

}
