% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_course.R
\name{init_course}
\alias{init_course}
\title{Initialize a lhCore course}
\usage{
init_course(
  course_proj,
  course_org = "nceas-learning-hub",
  course_title = NULL,
  course_desc = NULL,
  course_dates = NULL,
  setup_github = TRUE,
  template = c("lh", "adc", "delta", "corer")[1],
  package = "lhLessons",
  loc = ".",
  quiet = FALSE
)
}
\arguments{
\item{course_proj}{The name of the project directory (and Github repo) for the course.
Recommended format: YYYY-MM-\if{html}{\out{<course type>}}.}

\item{course_org}{The name of the Github organization to which the course
will belong.}

\item{course_title}{A brief but descriptive title for the course.  If NULL
(default), the title will be set the same as the project name.}

\item{course_desc}{A sentence or two describing the course.  If NULL
(default), will be left empty.}

\item{course_dates}{The dates of the course (as character, so any preferred format).
If NULL, will be left empty.}

\item{setup_github}{Set up the new course for Git/GitHub?  Default \code{TRUE} will
initialize the course repository as a Git-tracked project, and then will
connect it to a GitHub repository named \code{course_proj} in the organization
defined by \code{course_org}.}

\item{template}{The stylistic template to use when creating the course.  Template
defines the default landing/index page, banner colors, hex logo, etc.  Current
options are: \code{'lh'} (default) for Learning Hub style and landing page; \code{'adc'}
for Arctic Data Center; \code{'delta'} for Delta Stewardship Council; and \code{'corer'}
for coreR style and landing page.}

\item{package}{The lessons package to be associated with the course.  The lessons
package is the source of the Quarto structure and index page files.  Currently
just the one lessons package; this is a placeholder for potential future
functionality allowing for customized lessons packages for other users.}

\item{loc}{The file location where the course repository will be created.
Defaults to the current working directory.}

\item{quiet}{Provide progress and diagnostic messages during course initialization?
Default \code{FALSE}.}
}
\value{
stuff
}
\description{
This will set up the basic project structure for a new course: a project
repository, plus a file containing metadata about the course.  The metadata
file will be used to help populate details in later functions.
}
\examples{
\dontrun{
init_course(course_proj  = "2024-10-coreR",
            course_title = "CoreR October 2024",
            course_desc  = "CoreR course offered at NCEAS in Oct 2024",
            course_dates = "Oct. 6 - 10, 2024")
}

}
