% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_theme.R
\name{install_theme}
\alias{install_theme}
\title{Install a Quarto theme}
\usage{
install_theme(
  user = NULL,
  org = "nceas-learning-hub",
  repo = "lh_theme",
  theme = NULL
)
}
\arguments{
\item{user}{If theme is in a user repository, provide a user name (default \code{NULL}).  If \code{NULL},
the organization name will be used.  If a \code{user} repository is desired, set
\code{org} to \code{NULL} otherwise \code{org} will supercede \code{user}.}

\item{org}{If the theme is in an organization repository, provide the organization name.
Default: \code{nceas-learning-hub}.  If a \code{user} repo is desired, set \code{org} to \code{NULL}.}

\item{repo}{The repository name where the Quarto extension can be found.  Default:
Learning Hub theme \code{lh_theme}}

\item{theme}{The name of the theme within the \verb{_extensions} folder in the repository.
Defaults to the same name as the repo.}
}
\value{
If successful, returns the URL of the installed theme repository.
}
\description{
Provide information about a desired Quarto theme to be applied to the course site.
Defaults to nceas-learning-hub/lh_theme.
}
\examples{
\dontrun{}
}
