% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_lessons.R
\name{search_lessons}
\alias{search_lessons}
\title{Search available lessons using partial matching}
\usage{
search_lessons(query = NULL, pkg = "lhLessons", quiet = TRUE)
}
\arguments{
\item{query}{A character string to search across lessons in the given
package.  Regular expressions are allowed.  If no query given, returns all available lessons.}

\item{pkg}{The package to query for lesson availability (default \code{lhLessons})}
}
\value{
A data frame containing the file names (.qmd) of lessons
from the given package that match the search query.
}
\description{
Identify lessons whose file name matches a given search string.  Regular
expressions are allowed.
}
\examples{
\dontrun{search_lessons(query = "github"))}
}
