% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lessons.R
\name{setup_lessons}
\alias{setup_lessons}
\title{Set up a schedule of course lessons}
\usage{
setup_lessons(lessons, overwrite = FALSE)
}
\arguments{
\item{lessons}{A character vector containing the names (bare or .qmd extension) of the lessons, in order.}

\item{overwrite}{If there is already an existing lessons structure, and overwrite == FALSE, will return an error.  If the user wishes to create a new lesson structure, then overwrite == FALSE will remove the existing structure and create a new structure based on the lesson names.}
}
\description{
Set up a schedule of course lessons
}
\examples{
# setup_lessons('git-github-intro', 'r-creating-functions.qmd')
}
